+++ src/modals/NewAddressModal.tsx
@@ -26,12 +26,12 @@ import ColoredLabelInput from '../components/Inputs/ColoredLabelInput'
 import KeyValueInput from '../components/Inputs/InlineLabelValueInput'
 import Select from '../components/Inputs/Select'
 import Toggle from '../components/Inputs/Toggle'
-import ModalCentered, { ModalFooterButton, ModalFooterButtons } from '../components/ModalCentered'
 import HorizontalDivider from '../components/PageComponents/HorizontalDivider'
 import { Section } from '../components/PageComponents/PageContainers'
 import { Address, useAddressesContext } from '../contexts/addresses'
 import { useGlobalContext } from '../contexts/global'
 import { getRandomLabelColor } from '../utils/colors'
+import CenteredModal, { ModalFooterButton, ModalFooterButtons } from './CenteredModal'
 
 interface NewAddressModalProps {
   title: string
@@ -116,7 +116,7 @@ const NewAddressModal = ({ title, onClose, singleAddress }: NewAddressModalProps
   }
 
   return (
-    <ModalCentered title={title} onClose={onClose}>
+    <CenteredModal title={title} onClose={onClose}>
       <Section>
         <ColoredLabelInput placeholder="Address label" onChange={setAddressLabel} value={addressLabel} id="label" />
         {singleAddress && (
@@ -155,7 +155,7 @@ const NewAddressModal = ({ title, onClose, singleAddress }: NewAddressModalProps
         </ModalFooterButton>
         <ModalFooterButton onClick={onGenerateClick}>Generate</ModalFooterButton>
       </ModalFooterButtons>
-    </ModalCentered>
+    </CenteredModal>
   )
 }
