+++ src/modals/SendModal/CheckTransaction.tsx
@@ -20,7 +20,7 @@ import { convertAlphToSet, formatAmountForDisplay } from 'alephium-js'
 import styled from 'styled-components'
 
 import InfoBox from '../../components/InfoBox'
-import { ModalFooterButton, ModalFooterButtons } from '../../components/ModalCentered'
+import { ModalFooterButton, ModalFooterButtons } from '../CenteredModal'
 import { SendTransactionData } from '.'
 
 interface SendModalCheckTransactionProps {
